refactored Server to use JournalControllers moved JournalServices to frontend
